notification_service/admin/pyproject.toml
[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.ruff]
line-length = 120
select = ["D", "B", "S", "A", "COM", "C4", "T10", "F"]
ignore = [
    "D100",
    "D107",
    "D101",
    "D106",
    "D103",
    "D104",
    "D102",
    "D105",
    "B008",
    "B904",
    "A003",
]
exclude = ["__init__.py"]

[tool.poetry]
name = "admin"
version = "0.1.0"
description = ""
authors = ["likeinlife <likeinlife@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
celery = "5.2.7"
django-debug-toolbar = "3.4.0"
django-split-settings = "1.2.0"
django = "5.0"
flower = "2.0.1"
orjson = "3.9.10"
psycopg2-binary = "2.9.9"
pydantic = "2.5.3"
python-dotenv = "1.0.0"
redis = "4.5.1"
requests = "2.31.0"
uWSGI = "2.0.22"
dependency-injector = "^4.41.0"
pydantic-settings = "^2.1.0"
python-jose = "^3.3.0"
structlog = "^24.1.0"
django-structlog = "^7.1.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
ruff = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
